A100/Expert68/neurips_30b_submission/requirements.txt
transformers==4.32.0
bitsandbytes==0.41.1
accelerate==0.21.0
protobuf==4.23.4
sentencepiece==0.1.99
huggingface_hub
llama-cpp-python
